YoutubeAssistant/requirements.txt
langchain
openai
streamlit
python-dotenv
youtube-transcript-api
tiktoken
faiss-cpu